chore: Move animations symbols to Resources
