+++ src/front/macOS/kDriveResources/Symbols/Assets+Symbols.swift
