+++ src/front/macOS/kDriveResources/Symbols/Colors+Symbols.swift
