+++ src/front/macOS/kDriveResources/Symbols/Strings+Symbols.swift
