+++ src/front/macOS/swiftgen.yml
@@ -18,7 +18,7 @@ strings:
   inputs: kDriveResources/Localizable/en.lproj
   outputs:
     - templateName: structured-swift5
-      output: kDriveResources/Generated/Strings+Generated.swift
+      output: kDriveResources/Symbols/Strings+Symbols.swift
       params:
         enumName: KDriveLocalizable
         publicAccess: true
@@ -27,14 +27,14 @@ xcassets:
   - inputs: kDriveResources/Images.xcassets
     outputs:
       - templateName: swift5
-        output: kDriveResources/Generated/Assets+Generated.swift
+        output: kDriveResources/Symbols/Assets+Symbols.swift
         params:
           enumName: KDriveResources
           publicAccess: true
   - inputs: kDriveResources/Colors.xcassets
     outputs:
       - templateName: swift5
-        output: kDriveResources/Generated/Colors+Generated.swift
+        output: kDriveResources/Symbols/Colors+Symbols.swift
         params:
           enumName: KDriveColors
           publicAccess: true
